fixup! Convert dummy app to TS
